Deleted Wall, now Building has its own HP and Upgrade/Repair
