--- Assets/Scripts/Model/Wall.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: efb38b3ed4aabad4a802303481395f13
-timeCreated: 1476193254
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
